refactor(KlaviyoSwift): relocate state-store mirror off KlaviyoInternal (MAGE-832)
